Start of SlidingList re-implmentation
Used more sub-components and made use of tagging

There is the listManager to manage the former listOflists
listManager will attempt to find the catagory list and will check if it is
up to date. If either of these conditions is false it will create the list
CatagoryList is expected to have a categoryList tag
Catagory list is treated seperate from the actual catagories, paging will
involve only the catagories and the catagory list will be brought up by
colliding with the title box
Catagory scrolling will be handled in this component

Category components (needs renaming) to handle individual lists. It will
treat children with the "ListItem" tag as list items and automaticaly
posisition them.
Category component will handle it's own scrolling (but invoked from list
manager)