Started implmentation of interaction model for sliding list
There is a abstract SlidingListIteractionManager with relevent events that
sliding list can subscribe to.

The Concrete InteractionManager SlidingListCollisionInteractionManager takes a number of game objects as public fields that will serve as
triggers when they are collided with. A SlidingListIteractionOnCollision
component will be attached if it does not exist to each of the game
objects which will trigger a OnSelect event that will propogate down on
collsions